refactor: extracts engine
