+++ src/v2/Components/Inquiry/config.ts
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from "react"
-import { Engine } from "./Engine"
+import { Engine } from "v2/Utils/Engine"
 import { Context } from "./Hooks/useInquiryContext"
 import { Visited } from "./Visited"
 import { InquiryAccount } from "./Views/InquiryAccount"
